# openapi.yaml
openapi: 3.2.0
info:
  title: API REST de Pedidos
  version: 1.0.0
  description: API académica con validación, errores consistentes, consultas e idempotencia.
servers:
  - url: http://127.0.0.1:3000
tags:
  - name: Orders
paths:
  /orders:
    get:
      tags: [Orders]
      summary: Lista pedidos con filtros, ordenamiento y paginación
      parameters:
        - { $ref: '#/components/parameters/Status' }
        - { $ref: '#/components/parameters/CustomerName' }
        - { $ref: '#/components/parameters/Sort' }
        - { $ref: '#/components/parameters/OrderDirection' }
        - { $ref: '#/components/parameters/Page' }
        - { $ref: '#/components/parameters/PageSize' }
      responses:
        '200':
          description: Página de pedidos
          content:
            application/json:
              schema: { $ref: '#/components/schemas/OrderList' }
              example:
                data:
                  - id: 123e4567-e89b-12d3-a456-426614174000
                    customerName: Ana López
                    status: pending
                    items: [{ productId: P-100, quantity: 2, unitPrice: 25.5 }]
                    total: 51
                    createdAt: '2026-09-02T15:00:00.000Z'
                    updatedAt: '2026-09-02T15:00:00.000Z'
                pagination: { page: 1, pageSize: 10, totalItems: 1, totalPages: 1 }
        '400': { $ref: '#/components/responses/ValidationError' }
    post:
      tags: [Orders]
      summary: Crea un pedido de forma idempotente
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/CreateOrder' }
            example:
              customerName: Ana López
              items: [{ productId: P-100, quantity: 2, unitPrice: 25.5 }]
      responses:
        '201':
          description: Pedido creado por primera vez
          headers:
            Location:
              description: URI del pedido creado
              schema: { type: string, example: /orders/123e4567-e89b-12d3-a456-426614174000 }
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Order' }
        '200':
          description: Repetición con la misma clave y los mismos datos
          headers:
            Idempotency-Replayed:
              schema: { type: string, const: 'true' }
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Order' }
        '400': { $ref: '#/components/responses/ValidationError' }
        '409':
          description: Clave reutilizada con datos diferentes
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
              example:
                error:
                  code: IDEMPOTENCY_KEY_CONFLICT
                  message: La clave de idempotencia ya fue usada con datos diferentes
                  requestId: req-123
  /orders/{id}:
    parameters:
      - $ref: '#/components/parameters/OrderId'
    get:
      tags: [Orders]
      summary: Consulta un pedido
      responses:
        '200':
          description: Pedido encontrado
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Order' }
        '400': { $ref: '#/components/responses/ValidationError' }
        '404': { $ref: '#/components/responses/NotFound' }
    patch:
      tags: [Orders]
      summary: Actualiza parcialmente un pedido
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/UpdateOrder' }
            example: { status: confirmed }
      responses:
        '200':
          description: Pedido actualizado
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Order' }
        '400': { $ref: '#/components/responses/ValidationError' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409':
          description: El pedido ya está cancelado
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
    delete:
      tags: [Orders]
      summary: Cancela un pedido de forma idempotente
      description: Conserva el recurso y cambia su estado a cancelled. Repetir la solicitud mantiene el mismo resultado.
      responses:
        '204': { description: Pedido cancelado o ya cancelado }
        '400': { $ref: '#/components/responses/ValidationError' }
        '404': { $ref: '#/components/responses/NotFound' }
  /health:
    get:
      summary: Comprueba la disponibilidad
      responses:
        '200':
          description: Servicio disponible
          content:
            application/json:
              schema:
                type: object
                required: [status]
                properties: { status: { type: string, const: ok } }
  /openapi.yaml:
    get:
      summary: Devuelve este contrato OpenAPI
      responses:
        '200':
          description: Documento YAML
          content: { application/yaml: {} }
  /openapi.json:
    get:
      summary: Devuelve este contrato OpenAPI como JSON
      responses:
        '200':
          description: Documento JSON
          content: { application/json: {} }
components:
  parameters:
    OrderId:
      name: id
      in: path
      required: true
      schema: { type: string, format: uuid }
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: true
      description: Clave única de 8 a 128 caracteres. La misma clave y carga devuelve el pedido original.
      schema: { type: string, minLength: 8, maxLength: 128, pattern: '^[A-Za-z0-9._:-]+$' }
      example: demo-order-001
    Status:
      name: status
      in: query
      schema: { $ref: '#/components/schemas/OrderStatus' }
    CustomerName:
      name: customerName
      in: query
      description: Coincidencia parcial sin distinguir mayúsculas
      schema: { type: string, minLength: 1 }
    Sort:
      name: sort
      in: query
      schema: { type: string, enum: [createdAt, customerName, total], default: createdAt }
    OrderDirection:
      name: order
      in: query
      schema: { type: string, enum: [asc, desc], default: desc }
    Page:
      name: page
      in: query
      schema: { type: integer, minimum: 1, default: 1 }
    PageSize:
      name: pageSize
      in: query
      schema: { type: integer, minimum: 1, maximum: 50, default: 10 }
  responses:
    ValidationError:
      description: Parámetros, encabezados o contenido inválidos
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
          example:
            error:
              code: VALIDATION_ERROR
              message: La solicitud contiene datos inválidos
              details: [{ field: customerName, message: Too small }]
              requestId: req-123
    NotFound:
      description: Pedido no encontrado
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
          example:
            error: { code: ORDER_NOT_FOUND, message: Pedido no encontrado, requestId: req-123 }
  schemas:
    OrderStatus: { type: string, enum: [pending, confirmed, cancelled] }
    OrderItem:
      type: object
      additionalProperties: false
      required: [productId, quantity, unitPrice]
      properties:
        productId: { type: string, minLength: 1, maxLength: 50, example: P-100 }
        quantity: { type: integer, minimum: 1, maximum: 1000, example: 2 }
        unitPrice: { type: number, exclusiveMinimum: 0, maximum: 1000000, example: 25.5 }
    CreateOrder:
      type: object
      additionalProperties: false
      required: [customerName, items]
      properties:
        customerName: { type: string, minLength: 2, maxLength: 100 }
        items: { type: array, minItems: 1, maxItems: 100, items: { $ref: '#/components/schemas/OrderItem' } }
    UpdateOrder:
      type: object
      additionalProperties: false
      minProperties: 1
      properties:
        customerName: { type: string, minLength: 2, maxLength: 100 }
        status: { type: string, enum: [pending, confirmed] }
        items: { type: array, minItems: 1, maxItems: 100, items: { $ref: '#/components/schemas/OrderItem' } }
    Order:
      allOf:
        - $ref: '#/components/schemas/CreateOrder'
        - type: object
          required: [id, status, total, createdAt, updatedAt]
          properties:
            id: { type: string, format: uuid }
            status: { $ref: '#/components/schemas/OrderStatus' }
            total: { type: number, minimum: 0 }
            createdAt: { type: string, format: date-time }
            updatedAt: { type: string, format: date-time }
    Pagination:
      type: object
      required: [page, pageSize, totalItems, totalPages]
      properties:
        page: { type: integer, minimum: 1 }
        pageSize: { type: integer, minimum: 1 }
        totalItems: { type: integer, minimum: 0 }
        totalPages: { type: integer, minimum: 0 }
    OrderList:
      type: object
      required: [data, pagination]
      properties:
        data: { type: array, items: { $ref: '#/components/schemas/Order' } }
        pagination: { $ref: '#/components/schemas/Pagination' }
    ErrorDetail:
      type: object
      required: [field, message]
      properties: { field: { type: string }, message: { type: string } }
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message, requestId]
          properties:
            code: { type: string }
            message: { type: string }
            details: { type: array, items: { $ref: '#/components/schemas/ErrorDetail' } }
            requestId: { type: string }